Refactor concern, extract verify found records util into own trait
