+++ packages/Http/Api/src/Requests/Concerns/MultipleRecords.php
@@ -20,6 +20,8 @@ use Illuminate\Validation\ValidationException;
  */
 trait MultipleRecords
 {
+    use RecordExistence;
+
     /**
      * List of requested target records
      *
@@ -197,47 +199,6 @@ trait MultipleRecords
     }
 
     /**
-     * Verifies that all requested records are found or fails
-     *
-     * @param string[]|int[] $requested List of unique identifiers
-     * @param Collection<Model> $found Collection of found records
-     * @param string $matchKey Key to match from found records
-     * @param string $targetsKey Name of property in request payload that contains "targets"
-     *
-     * @return Collection<Model>
-     *
-     * @throws ValidationException
-     */
-    public function verifyAllRecordsFound(
-        array $requested,
-        Collection $found,
-        string $matchKey,
-        string $targetsKey
-    ): Collection {
-        // When the amount found matches amount requested, then we assume that all were
-        // found. Might not be the most correct, but should be fast...
-        if ($found->count() === count($requested)) {
-            return $found;
-        }
-
-        // Otherwise, the difference must be identified and exception thrown.
-        $foundValues = $found->pluck($matchKey)->toArray();
-        $difference = array_diff($requested, $foundValues);
-
-        $errors = [];
-        foreach ($difference as $notFound) {
-            $index = array_search($notFound, $requested);
-            if ($index === false) {
-                $index = $notFound;
-            }
-
-            $errors["{$targetsKey}.{$index}"] = $this->makeRecordNotFoundMessage($notFound, $index);
-        }
-
-        throw ValidationException::withMessages($errors);
-    }
-
-    /**
      * Accepts integer values for {@see targetsKey()} property
      *
      * @param string $key [optional] Name of unique key in model
@@ -399,20 +360,5 @@ trait MultipleRecords
     {
         return $query
             ->whereIn($key, $targets);
-    }
-
-    /**
-     * Returns a "record not found" error message
-     *
-     * @param string|int $target
-     * @param string|int $index
-     *
-     * @return string
-     */
-    protected function makeRecordNotFoundMessage(string|int $target, string|int $index): string
-    {
-        $key = 'athenaeum-http-api::api-resources.record_not_found';
-
-        return Lang::get($key, [ 'record' => $target ]);
     }
 }
